+++ lib/cocoapods/installer/xcode.rb
@@ -2,6 +2,7 @@ module Pod
   class Installer
     class Xcode
       autoload :PodsProjectGenerator, 'cocoapods/installer/xcode/pods_project_generator'
+      autoload :TargetValidator, 'cocoapods/installer/xcode/target_validator'
     end
   end
 end